requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pymongo==4.0.2
six==1.16.0
Werkzeug==2.0.3
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.24.0
cryptography==3.4.7